feat:채팅방 생성, 채팅메세지 전송, 채팅방 참가 구현, 테스트코드 구현
fix: repository에서는 Long -> String으로 수정
domian -> entity로 수정
