+++ src/main/java/com/team1/epilogue/chat/repository/ParticipationRepository.java
@@ -1,10 +1,10 @@
 package com.team1.epilogue.chat.repository;
 
-import com.team1.epilogue.chat.domain.Participation;
+import com.team1.epilogue.chat.entity.Participation;
 import org.springframework.data.mongodb.repository.ReactiveMongoRepository;
 import reactor.core.publisher.Flux;
 
-public interface ParticipationRepository extends ReactiveMongoRepository<Participation,Long> {
+public interface ParticipationRepository extends ReactiveMongoRepository<Participation,String> {
   //특정 채팅방(roomId)에 참여한 모든 사용자 조회
   Flux<Participation> findByRoomId(String roomId);
