+++ src/main/java/com/team1/epilogue/chat/service/ChatMessageService.java
@@ -1,6 +1,6 @@
 package com.team1.epilogue.chat.service;
 
-import com.team1.epilogue.chat.domain.ChatMessage;
+import com.team1.epilogue.chat.entity.ChatMessage;
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
