+++ src/main/java/com/team1/epilogue/chat/service/Impl/ChatMessageServiceImpl.java
@@ -1,6 +1,6 @@
 package com.team1.epilogue.chat.service.Impl;
 
-import com.team1.epilogue.chat.domain.ChatMessage;
+import com.team1.epilogue.chat.entity.ChatMessage;
 import com.team1.epilogue.chat.repository.ChatMessageRepository;
 import com.team1.epilogue.chat.service.ChatMessageService;
 import java.time.LocalDateTime;
